refactored the html
